Test now passes in Firefox 3.6.22
Firefox generated multiple errors instead of just one. Since response.js always tried to remove <script id="script0" .../> this lead to another error being raised, jsErrLog catching it, and then infinite recursion.

The reason why the increment of err_i was moved earlier was to eliminate the race condition that might occur in the testing environment when removeScript is called in response.js. It does not have any effect in production.

The test has also been verified to work on Google Chrome 14.0.835.163.